Change function name from getLabelByTag to getBylabel and  remove CentreonData file.
