+++ tests/e2e/cypress/integration/Resources-status/common.ts
@@ -7,8 +7,8 @@ import {
   submitResultsViaClapi,
   versionApi,
   logout,
+  insertFixture,
 } from '../../commons';
-import { initializeResourceData } from '../../support/centreonData';
 
 interface Criteria {
   name: string;
@@ -28,6 +28,26 @@ const refreshButton = '[aria-label="Refresh"]';
 const resourceMonitoringApi = /.+api\/beta\/monitoring\/resources.?page.+/;
 
 const apiFilterResources = `${apiBase}/${versionApi}/users/filters/events-view`;
+
+const initializeResourceData = (): Cypress.Chainable => {
+  const files = [
+    'resources/clapi/host1/01-add.json',
+    'resources/clapi/service1/01-add.json',
+    'resources/clapi/service1/02-set-max-check.json',
+    'resources/clapi/service1/03-disable-active-check.json',
+    'resources/clapi/service1/04-enable-passive-check.json',
+    'resources/clapi/service2/01-add.json',
+    'resources/clapi/service2/02-set-max-check.json',
+    'resources/clapi/service2/03-disable-active-check.json',
+    'resources/clapi/service2/04-enable-passive-check.json',
+    'resources/clapi/service3/01-add.json',
+    'resources/clapi/service3/02-set-max-check.json',
+    'resources/clapi/service3/03-disable-active-check.json',
+    'resources/clapi/service3/04-enable-passive-check.json',
+  ];
+
+  return cy.wrap(Promise.all(files.map(insertFixture)));
+};
 
 const insertResourceFixtures = (): Cypress.Chainable => {
   return loginAsAdminViaApiV2()
